+++ Library/Aphid/Model.js
@@ -127,6 +127,8 @@
  *
  * ### Callback Methods
  *
+ *  * `afterLoad()` - Called immediately after a successful load operation.
+ *
  *  * `afterSave()` - Called immediately after a successful save operation.
  *
  * ### Delegate Methods
@@ -324,8 +326,7 @@ Aphid.Model = Class.create({
         this.object = transport.responseJSON;
         this._initializeFromObject();
         this.isLoaded = true;
-        if (this.delegate && this.delegate.modelDidFinishLoading)
-          this.delegate.modelDidFinishLoading(this);
+        this._afterLoad();
       }.bind(this),
       onFailure: function(transport)
       {
@@ -527,6 +528,14 @@ Aphid.Model = Class.create({
   },
 
   // Callbacks ---------------------------------------------------------------
+
+  _afterLoad: function()
+  {
+    if (this.afterLoad)
+      this.afterLoad();
+    if (this.delegate && this.delegate.modelDidFinishLoading)
+      this.delegate.modelDidFinishLoading(this);
+  },
 
   _afterSave: function()
   {